[KP-19]Deleted Scene Delegate
Now project starts in AppDelegate
